+++ OmmelSamvirke.Domain/Features/Newsletters/Models/NewsletterSubscription.cs
@@ -1,6 +1,0 @@
-﻿namespace OmmelSamvirke.Domain.Features.Newsletters.Models;
-
-public class NewsletterSubscription
-{
-    
-}